Removed duplicate study.
There were two entries for "A Two-Year Randomized Weight Loss Trial Comparing a Vegan Diet to a More Moderate Low-Fat Diet". They had different sources as one was full text and the other abstract only, but I don't think there's a reason to keep the abstract only one. Maybe we should add DOI as a unique key to prevent this in future, but that is probably unnecessary & overkill.